refactor(model): working model package
We refactored the code in order to be able to use the package correctly
