+++ model/.idea/inspectionProfiles/Project_Default.xml
@@ -1,0 +1,71 @@
+<component name="InspectionProjectProfileManager">
+  <profile version="1.0">
+    <option name="myName" value="Project Default" />
+    <inspection_tool class="Eslint" enabled="true" level="WARNING" enabled_by_default="true" />
+    <inspection_tool class="PyPackageRequirementsInspection" enabled="true" level="WARNING" enabled_by_default="true">
+      <option name="ignoredPackages">
+        <value>
+          <list size="57">
+            <item index="0" class="java.lang.String" itemvalue="basemap" />
+            <item index="1" class="java.lang.String" itemvalue="opencv-python" />
+            <item index="2" class="java.lang.String" itemvalue="numpy" />
+            <item index="3" class="java.lang.String" itemvalue="absl-py" />
+            <item index="4" class="java.lang.String" itemvalue="google-pasta" />
+            <item index="5" class="java.lang.String" itemvalue="mlopspython-extraction" />
+            <item index="6" class="java.lang.String" itemvalue="protobuf" />
+            <item index="7" class="java.lang.String" itemvalue="opt-einsum" />
+            <item index="8" class="java.lang.String" itemvalue="h11" />
+            <item index="9" class="java.lang.String" itemvalue="pymupdf" />
+            <item index="10" class="java.lang.String" itemvalue="setuptools" />
+            <item index="11" class="java.lang.String" itemvalue="gast" />
+            <item index="12" class="java.lang.String" itemvalue="markdown-it-py" />
+            <item index="13" class="java.lang.String" itemvalue="markupsafe" />
+            <item index="14" class="java.lang.String" itemvalue="starlette" />
+            <item index="15" class="java.lang.String" itemvalue="certifi" />
+            <item index="16" class="java.lang.String" itemvalue="anyio" />
+            <item index="17" class="java.lang.String" itemvalue="astunparse" />
+            <item index="18" class="java.lang.String" itemvalue="keras" />
+            <item index="19" class="java.lang.String" itemvalue="uvicorn" />
+            <item index="20" class="java.lang.String" itemvalue="mlopspython-inference" />
+            <item index="21" class="java.lang.String" itemvalue="libclang" />
+            <item index="22" class="java.lang.String" itemvalue="pydantic" />
+            <item index="23" class="java.lang.String" itemvalue="werkzeug" />
+            <item index="24" class="java.lang.String" itemvalue="h5py" />
+            <item index="25" class="java.lang.String" itemvalue="wrapt" />
+            <item index="26" class="java.lang.String" itemvalue="tensorboard-data-server" />
+            <item index="27" class="java.lang.String" itemvalue="typing-extensions" />
+            <item index="28" class="java.lang.String" itemvalue="click" />
+            <item index="29" class="java.lang.String" itemvalue="flatbuffers" />
+            <item index="30" class="java.lang.String" itemvalue="tensorboard" />
+            <item index="31" class="java.lang.String" itemvalue="optree" />
+            <item index="32" class="java.lang.String" itemvalue="charset-normalizer" />
+            <item index="33" class="java.lang.String" itemvalue="idna" />
+            <item index="34" class="java.lang.String" itemvalue="requests" />
+            <item index="35" class="java.lang.String" itemvalue="sniffio" />
+            <item index="36" class="java.lang.String" itemvalue="exceptiongroup" />
+            <item index="37" class="java.lang.String" itemvalue="tensorflow" />
+            <item index="38" class="java.lang.String" itemvalue="mdurl" />
+            <item index="39" class="java.lang.String" itemvalue="markdown" />
+            <item index="40" class="java.lang.String" itemvalue="ml-dtypes" />
+            <item index="41" class="java.lang.String" itemvalue="pymupdfb" />
+            <item index="42" class="java.lang.String" itemvalue="urllib3" />
+            <item index="43" class="java.lang.String" itemvalue="pygments" />
+            <item index="44" class="java.lang.String" itemvalue="coverage" />
+            <item index="45" class="java.lang.String" itemvalue="six" />
+            <item index="46" class="java.lang.String" itemvalue="wheel" />
+            <item index="47" class="java.lang.String" itemvalue="rich" />
+            <item index="48" class="java.lang.String" itemvalue="packaging" />
+            <item index="49" class="java.lang.String" itemvalue="python-multipart" />
+            <item index="50" class="java.lang.String" itemvalue="tensorflow-io-gcs-filesystem" />
+            <item index="51" class="java.lang.String" itemvalue="termcolor" />
+            <item index="52" class="java.lang.String" itemvalue="fastapi" />
+            <item index="53" class="java.lang.String" itemvalue="opencv-contrib-python-headless" />
+            <item index="54" class="java.lang.String" itemvalue="namex" />
+            <item index="55" class="java.lang.String" itemvalue="pillow" />
+            <item index="56" class="java.lang.String" itemvalue="grpcio" />
+          </list>
+        </value>
+      </option>
+    </inspection_tool>
+  </profile>
+</component>
